refactor(vehicle-view-component): add photos tab component
* Add photos tab component and move all photos logics from vehicle-view component into photos tab component.
* Rename the selector of vehicle-view-basics-tab-content component to make it more related to its role.
